requirements.txt
Django==5.2.7
PyPDF2==3.0.1
python-docx==1.2.0
openai==2.6.1
